2024-02-update-1
CHANGED. Cleaned up abjad.Dynamic.

    CHANGED: abjad.Dynamic is now a frozen dataclass

    CHANGED: added abjad.Dynamic.get_ordinal()

    REMOVED: abjad.Dynamic("niente").

        * OLD: abjad.Dynamic("niente")
        * NEW: abjad.StopHaipin()

    Moved abjad.Dynamic to indicators.py.

FIXED. #1577:

    OLD:

        abjad.iterpitches.transpose_from_sounding_pitch() lost note head tweaks
        abjad.iterpitches.transpose_from_written_pitch() lost note head tweaks

    NEW:

        abjad.iterpitches.transpose_from_sounding_pitch() preserves note head tweaks
        abjad.iterpitches.transpose_from_written_pitch() preserves note head tweaks

NEW Added abjad.StartTrillSpan.force_trill_pitch_head_accidental keyword.

    * Set to true to force "!" after pitch name.
