+++ abjad/math.py
@@ -1,6 +1,7 @@
 """
 Abjad's math library.
 """
+
 import collections
 import fractions
 import itertools
